+++ crates/action/src/action_node.rs
@@ -30,6 +30,7 @@ pub struct RunTaskNode {
     pub persistent: bool,  // Never terminates
     pub runtime: Runtime,
     pub target: Target,
+    pub timeout: Option<u64>,
 }
 
 impl RunTaskNode {
@@ -41,6 +42,7 @@ impl RunTaskNode {
             persistent: false,
             runtime,
             target,
+            timeout: None,
         }
     }
 }