+++ packages/report/src/action.ts
@@ -11,6 +11,8 @@ export function getIconForStatus(status: ActionStatus): string {
 			return '🟦';
 		case 'failed':
 		case 'failed-and-abort':
+		case 'aborted':
+		case 'timed-out':
 			return '🟥';
 		case 'invalid':
 			return '🟨';
@@ -25,7 +27,12 @@ export function getIconForStatus(status: ActionStatus): string {
 }
 
 export function hasFailed(status: ActionStatus): boolean {
-	return status === 'failed' || status === 'failed-and-abort';
+	return (
+		status === 'failed' ||
+		status === 'failed-and-abort' ||
+		status === 'aborted' ||
+		status === 'timed-out'
+	);
 }
 
 export function hasPassed(status: ActionStatus): boolean {