+++ packages/cli/CHANGELOG.md
@@ -9,7 +9,7 @@
   - Increased performance.
   - Better concurrency handling and scheduling.
   - More accurately monitors signals (ctrl+c) and shutdowns.
-  - Tasks can now be configured with a timeout.
+  - Tasks can now be configured with a timeout (`options.timeout` setting).
 - Added a global `--dump` flag, that will dump a trace profile that can be inspected in
   `chrome://tracing`.
 - Updated `moon completions` command to support Nushell.